cape/parrot: Setup ramoops sizes
QCOM only sets the PMSG size, this region is used
only for storing logs from userspace.

Instead split the allocated region of 0x200000 bytes
into same sized chunks for each supported region by
the minidump driver that actually ends up preserving
the memory.

record-size is the size of chunk data to write in case
of dmesg dumping, the value was more or less randomly
picked, as long as it is a multiple of 4 it'll work.

The size used for dmesg is calculated from
removing the other reseverd memory sizes from
the total available memory region.

We now have support for DMESG, CONSOLE, FTRACE and PMSG

Change-Id: I7ff6be47f3d2e78d6ab6f1b62f75e9aba63bc337
Signed-off-by: Luca Stefani <luca@osomprivacy.com>